+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Treasure bags can now be used as ingredients in recipes. See wiki for details.
 - Shaped and shapeless recipe types for treasure bags. This allows you to create recipes which craft specific treasure bags. You can, of course, combine this with treasure bag ingredients as well.
 - JEI can distinguish treasure bags of different types
+### Changed
+- Treasure bags now display the data pack's ID in the tooltip. Unknown/invalid bags are also indicated in the tooltip.
 ### Fixed
 - Bag types should now synchronize with connecting clients on both dedicated servers and LAN games